[bug-fix] correct sub-sec
git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_5_0_exp@2908 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/utils/timestamp.h
@@ -37,7 +37,7 @@ public:
   }
   Timestamp (struct timeval *tv)
   {
-    ts = tv->tv_sec + tv->tv_usec / USEC_SEC;
+    ts = tv->tv_sec + (double) tv->tv_usec / USEC_SEC;
   }
   void setTs (double _ts)
   {
